% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_final_standing.R
\name{calculate_final_standing}
\alias{calculate_final_standing}
\title{Calculates league table based on result}
\usage{
calculate_final_standing(home, away, home_goals, away_goals)
}
\arguments{
\item{home}{vector of home teams}

\item{away}{vector of away teams}

\item{home_goals}{vector of home team goals}

\item{away_goals}{vector of away team goals}
}
\value{
final table
}
\description{
Calculates league table based on result
}
