% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-utils.R
\docType{data}
\name{CONFIG}
\alias{CONFIG}
\title{Package Configuration Settings}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
CONFIG
}
\description{
Configuration object containing league settings, forecast parameters, and file paths.
}
\keyword{datasets}
