% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-utils.R
\name{create_dir_if_missing}
\alias{create_dir_if_missing}
\title{Create directory if it doesn't exist}
\usage{
create_dir_if_missing(path)
}
\arguments{
\item{path}{Directory path to create}
}
\description{
Create directory if it doesn't exist
}
