% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-utils.R
\name{get_forecast_path}
\alias{get_forecast_path}
\title{Get path to forecast file}
\usage{
get_forecast_path(year, league_id, matchday)
}
\arguments{
\item{year}{Season end year}

\item{league_id}{League ID}

\item{matchday}{Matchday number}
}
\value{
Full path to forecast file
}
\description{
Get path to forecast file
}
