% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-utils.R
\name{get_mv_path}
\alias{get_mv_path}
\title{Get path to market value data file}
\usage{
get_mv_path(league_id, year)
}
\arguments{
\item{league_id}{League ID}

\item{year}{Season end year}
}
\value{
Full path to market value file
}
\description{
Get path to market value data file
}
