% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-utils.R
\name{get_results_path}
\alias{get_results_path}
\title{Get path to results file}
\usage{
get_results_path(country, level, year)
}
\arguments{
\item{country}{Country code (e.g., "FRA", "GER")}

\item{level}{League level (1, 2, 3, etc.)}

\item{year}{Season end year}
}
\value{
Full path to results file
}
\description{
Get path to results file
}
