% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-utils.R
\name{get_strength_pred_path}
\alias{get_strength_pred_path}
\title{Get path to strength prediction file}
\usage{
get_strength_pred_path(year, league_id)
}
\arguments{
\item{year}{Season end year}

\item{league_id}{League ID}
}
\value{
Full path to strength prediction file
}
\description{
Get path to strength prediction file
}
