% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_leagues}
\alias{load_leagues}
\title{Load Leagues}
\usage{
load_leagues(from_app = FALSE)
}
\arguments{
\item{from_app}{Logical indicating whether function is called from shiny app,
in which case the working directory is different, and ".." needs to be
appended to path.}
}
\value{
Data frame leagues data.
}
\description{
Loads file with info on leagues.
}
