% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpoptimize.R
\name{objective_fn}
\alias{objective_fn}
\title{Objective Function Wrapper for Hyperparameter Scoring}
\usage{
objective_fn(hp, n_sims, years, league_ids, parallel)
}
\arguments{
\item{hp}{A list of hyperparameter values to evaluate.}

\item{n_sims}{Number of simulations to run for evaluation.}

\item{years}{Vector of years to include in the evaluation.}

\item{league_ids}{Vector of league identifiers to use for evaluation.}

\item{parallel}{Logical indicating whether to use parallel processing.}
}
\value{
Numeric score from score_hyperparameters (lower is better).
}
\description{
Internal function that wraps the score_hyperparameters function call
with the appropriate arguments structure for use within hpoptimize.
}
\details{
This function serves as an adapter between the optimization framework
and the score_hyperparameters function. It constructs the appropriate
argument list and handles the function call.
}
\note{
This is an internal function used by hpoptimize and typically
should not be called directly by users.
}
\keyword{internal}
