% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpoptimize.R
\name{save_round_configs}
\alias{save_round_configs}
\title{Save round configurations}
\usage{
save_round_configs(checkpoint_dir, round, configs)
}
\arguments{
\item{checkpoint_dir}{Directory for checkpoints}

\item{round}{Round number}

\item{configs}{Configurations to save}
}
\description{
Internal helper function to save configurations for a round
}
\keyword{internal}
