% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpoptimize.R
\name{save_round_scores}
\alias{save_round_scores}
\title{Save round scores}
\usage{
save_round_scores(checkpoint_dir, round, scores)
}
\arguments{
\item{checkpoint_dir}{Directory for checkpoints}

\item{round}{Round number}

\item{scores}{Scores to save}
}
\description{
Internal helper function to save scores for a round
}
\keyword{internal}
