% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpoptimize.R
\name{write_log}
\alias{write_log}
\title{Write to log file}
\usage{
write_log(log_file, message, append = TRUE)
}
\arguments{
\item{log_file}{Path to the log file}

\item{message}{Message to write}

\item{append}{Whether to append to existing file}
}
\description{
Internal helper function to write messages to the log file
}
\keyword{internal}
